services/cartorders/requirements.txt
Faker==30.8.2
amqp==5.3.1
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.0
astroid==4.0.4
Authlib==1.6.7
bandit==1.9.3
billiard==4.2.4
black==26.1.0
celery==5.6.2
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
coverage==7.13.3
cryptography==46.0.4
dill==0.4.1
dj-database-url==3.1.0
Django==6.0.1
django-cors-headers==4.9.0
django-redis==6.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dparse==0.6.4
drf-yasg==1.21.14
django-filter==25.1
filelock==3.20.3
flake8==7.3.0
gunicorn==25.0.3
h11==0.16.0
hiredis==3.3.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
inflection==0.5.1
iniconfig==2.3.0
isort==7.0.0
Jinja2==3.1.6
joblib==1.5.3
kombu==5.6.2
mando==0.7.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==4.2.2
mccabe==0.7.0
mdurl==0.1.2
mypy_extensions==1.1.0
nltk==3.9.2
packaging==26.0
pathspec==1.0.4
pillow==12.1.0
platformdirs==4.5.1
pluggy==1.6.0
postgres==4.0
prompt_toolkit==3.0.52
psycopg==3.3.2
psycopg2-binary==2.9.11
psycopg2-pool==1.2
pycodestyle==2.14.0
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.10.1
pylint==4.0.4
pylint-django==2.7.0
pylint-plugin-utils==0.9.0
pytest==9.0.2
pytest-cov==7.0.0
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytokens==0.4.1
pytz==2025.2
PyYAML==6.0.3
radon==6.0.1
redis==7.1.0
regex==2026.1.15
reportlab==4.4.9
requests==2.32.5
rich==14.3.2
ruamel.yaml==0.19.1
safety==3.7.0
safety-schemas==0.0.16
shellingham==1.5.4
six==1.17.0
sqlparse==0.5.5
stevedore==5.6.0
tenacity==9.1.4
tomlkit==0.14.0
tqdm==4.67.3
typer==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
whitenoise==6.11.0
